feat:删除无用任务，新增霍夫曼编码。

+++ src/main/java/com/wang/Leetcode/weekContest/week319/No2471.java
@@ -20,7 +20,7 @@ public class No2471 {
 
             while (count>0){
                 TreeNode temp = queue.poll();
-                arr.add(temp.val);
+                arr.add((int)temp.val);
                 if (temp.left!=null){
                     queue.offer(temp.left);
                 }
